+++ src/pointbuffers.c
@@ -7,7 +7,7 @@
 
 static inline double rand_uniform(double low, double high);
 
-PointBuffer *ptbuf_new(int size) {
+PointBuffer *ptbuf_new(long size) {
     if (size < 1) {
         EMIT_ERR_MSG(ValueError, "PointBuffer size must be positive");
         return NULL;
@@ -33,7 +33,7 @@ PointBuffer *ptbuf_new(int size) {
 }
 
 PointBuffer *ptbuf_from_buffers(
-    int size,
+    long size,
     double *x_vals,
     double *y_vals,
     double *z_vals
@@ -43,14 +43,14 @@ PointBuffer *ptbuf_from_buffers(
         return NULL;
     }
     Point *points = buffer->points;
-    for (int i = 0; i < size; ++i) {
+    for (long i = 0; i < size; ++i) {
         points[i] = point_create(x_vals[i], y_vals[i], z_vals[i]);
     }
     return buffer;
 }
 
 PointBuffer *ptbuf_from_buffers_weighted(
-    int size,
+    long size,
     double *x_vals,
     double *y_vals,
     double *z_vals,
@@ -61,7 +61,7 @@ PointBuffer *ptbuf_from_buffers_weighted(
         return NULL;
     }
     Point *points = buffer->points;
-    for (int i = 0; i < size; ++i) {
+    for (long i = 0; i < size; ++i) {
         points[i].weight = weights[i];
     }
     return buffer;
@@ -74,7 +74,7 @@ void ptbuf_free(PointBuffer *buffer) {
     free(buffer);
 }
 
-int ptbuf_resize(PointBuffer *buffer, int size) {
+int ptbuf_resize(PointBuffer *buffer, long size) {
     if (size < 1) {
         EMIT_ERR_MSG(ValueError, "PointBuffer size must be positive");
         return BTR_FAILED;
@@ -108,13 +108,13 @@ static inline double rand_uniform(double low, double high) {
     return rand_uniform_normalised * (high - low) + low;
 }
 
-PointBuffer *ptbuf_gen_random(double low, double high, int num_points) {
+PointBuffer *ptbuf_gen_random(double low, double high, long num_points) {
     PointBuffer *buffer = ptbuf_new(num_points);
     if (buffer == NULL) {
         return NULL;
     }
 
-    for (int i = 0; i < num_points; ++i) {
+    for (long i = 0; i < num_points; ++i) {
         double x = rand_uniform(low, high);
         double y = rand_uniform(low, high);
         double z = rand_uniform(low, high);
@@ -135,6 +135,6 @@ PointSlice *ptslc_from_buffer(const PointBuffer *buffer) {
     return slice;
 }
 
-int ptslc_get_size(const PointSlice *slice) {
+long ptslc_get_size(const PointSlice *slice) {
     return slice->end - slice->start;
 }
